adk_example/google-multi-agent/pyproject.toml
[project]
name = "google-multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [ 
    "google-adk>=1.0.0",
    "google-genai>=1.11.0"
    ]
